splitting code
