+++ py/web/main.py
@@ -19,11 +19,11 @@ from .middleware import csrf_middleware, error_middleware, pg_middleware, user_m
 from .views import index, sitemap
 from .views.auth import (authenticate_token, guest_signup, host_signup, login, login_with, logout,
                          reset_password_request, set_password, unsubscribe)
-from .views.booking import (BookFreeTickets, BuyTickets, CancelReservedTickets, Donate, DonationOptionBread,
-                            ReserveTickets, booking_info)
+from .views.booking import BookFreeTickets, BuyTickets, CancelReservedTickets, ReserveTickets, booking_info
 from .views.categories import (CategoryBread, category_add_image, category_default_image, category_delete_image,
                                category_images, category_public, donation_options)
 from .views.company import CompanyBread, company_upload
+from .views.donate import Donate, DonationOptionBread
 from .views.emails import clear_email_def, email_def_browse, email_def_edit, email_def_retrieve
 from .views.events import (EventBread, EventUpdate, SetEventStatus, SetTicketTypes, event_categories, event_get,
                            event_ticket_types, event_tickets, event_tickets_export, event_updates_sent,
