+++ py/web/views/booking.py
@@ -3,16 +3,14 @@ from time import time
 from typing import List
 
 from asyncpg import CheckViolationError
-from buildpg import MultipleValues, V, Values
+from buildpg import MultipleValues, Values
 from buildpg.asyncpg import BuildPgConnection
-from buildpg.clauses import Join, Where
-from pydantic import BaseModel, EmailStr, condecimal, constr, validator
+from pydantic import BaseModel, EmailStr, constr, validator
 
 from web.actions import ActionTypes, record_action, record_action_id
 from web.auth import check_grecaptcha, check_session, is_auth
-from web.bread import Bread, UpdateView
-from web.stripe import (BookingModel, Reservation, StripeBuyModel, StripeDonateModel, book_free, stripe_buy,
-                        stripe_donate)
+from web.bread import UpdateView
+from web.stripe import BookingModel, Reservation, StripeBuyModel, book_free, stripe_buy
 from web.utils import JsonErrors, decrypt_json, encrypt_json, json_response
 
 logger = logging.getLogger('nosht.booking')
@@ -223,57 +221,3 @@ class BookFreeTickets(UpdateView):
         booked_action_id = await book_free(m, self.request['company_id'], self.session.get('user_id'),
                                            self.app, self.conn)
         await self.app['email_actor'].send_event_conf(booked_action_id)
-
-
-class Donate(UpdateViewAuth):
-    Model = StripeDonateModel
-
-    async def execute(self, m: StripeDonateModel):
-        await check_grecaptcha(m, self.request)
-        action_id, _ = await stripe_donate(m, self.request['company_id'], self.session['user_id'], self.app, self.conn)
-        await self.app['email_actor'].send_donation_thanks(action_id)
-
-
-class DonationOptionBread(Bread):
-    class Model(BaseModel):
-        category: int
-        name: constr(max_length=255)
-        amount: condecimal(ge=1, max_digits=6, decimal_places=2)
-        live: bool = True
-        short_description: str
-        long_description: str
-        sort_index: int = None
-
-    browse_enabled = True
-    retrieve_enabled = True
-    edit_enabled = True
-    add_enabled = True
-    delete_enabled = True
-
-    model = Model
-    table = 'donation_options'
-    table_as = 'opt'
-    browse_order_by_fields = 'opt.category', 'opt.sort_index', 'opt.amount'
-    browse_fields = (
-        'opt.id',
-        'opt.name',
-        V('cat.name').as_('category_name'),
-        'opt.live',
-        'opt.amount',
-    )
-    retrieve_fields = browse_fields + (
-        'opt.category',
-        'opt.sort_index',
-        'opt.short_description',
-        'opt.long_description',
-        'opt.image',
-    )
-
-    async def check_permissions(self, method):
-        await check_session(self.request, 'admin')
-
-    def where(self):
-        return Where(V('cat.company') == self.request['company_id'])
-
-    def join(self):
-        return Join(V('categories').as_('cat').on(V('cat.id') == V('opt.category')))
